requirements.txt
# After pip install -r requirements.txt, run: python -m spacy download en_core_web_sm
#
# Heavy local ML models (transformers / torch / biogpt) are NO LONGER needed.
# All NLP is handled by the Gemini API (google-generativeai).

google-genai>=1.0.0
spacy>=3.7.0
pdfplumber>=0.10.0
python-docx>=1.1.0
openai>=1.0.0
tqdm>=4.66.0
